stt-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
torch>=2.1.0
transformers>=4.36.0
librosa>=0.10.1
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
soundfile>=0.12.1
py_eureka_client>=0.11.0
google-generativeai>=0.3.0
